doc: `generate metrics` should include metrics
Created by builtin plugins: Aggregator, Max, and MetricAgent
